:construction: Fix: util migration(JS to TS)
Fix util migration(JS to TS)
