+++ src/components/common/Card.js
@@ -2,46 +2,13 @@ import styles from './Card.module.css';
 import classNames from 'classnames/bind';
 const cn = classNames.bind(styles);
 
+import { createDate, intervalDate } from '../../utils/Date';
+
 import noneData from '../../assets/images/none-data.png';
 import iconStar from '../../assets/icons/star-off.svg';
 import iconKebab from '../../assets/icons/kebab.svg';
 
 import CardPopover from '../FolderPage/CardPopover';
-
-const createDate = (value) => {
-  const date = new Date(value);
-  return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
-};
-
-const intervalDate = (value) => {
-  const createdDate = new Date(value);
-  const currentDate = new Date();
-
-  const diffMinute = Math.floor((currentDate - createdDate) / 1000 / 60);
-  const diffHour = Math.floor(diffMinute / 60);
-  const diffDay = Math.floor(diffHour / 24);
-  const diffMonth = Math.floor(diffDay / 30);
-  const diffYear = Math.floor(diffMonth / 12);
-
-  if (diffMinute < 2) {
-    return '1 minute ago';
-  }
-  if (diffMinute <= 59) {
-    return `${diffMinute} minutes ago`;
-  }
-  if (diffHour <= 23) {
-    return `${diffHour} hour${diffHour < 2 ? '' : 's'} ago`;
-  }
-  if (diffDay <= 30) {
-    return `${diffDay} day${diffDay < 2 ? '' : 's'} ago`;
-  }
-  if (diffMonth <= 11) {
-    return `${diffMonth} month${diffMonth < 2 ? '' : 's'} ago`;
-  }
-  else {
-    return `${diffYear} year${diffYear < 2 ? '' : 's'} ago`;
-  }
-};
 
 const Card = ({ card, isIconVisible = true }) => {
   const { url, title, description, createdAt, imageSource } = card;
